adhd_assistant/requirements.txt
python-telegram-bot[job-queue]>=21.0
google-generativeai>=0.8.0
aiosqlite>=0.20.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
